requirements.txt
Django
django-mailer
djangorestframework
django-cors-headers
whitenoise
